refactor(editor): move cloudPlan.store to @n8n/stores (no-changelog)
Move cloudPlan.store from editor-ui into the @n8n/stores package,
following the settings.store precedent (N8N-32 / #34574).

Path B: instead of dragging the app-only hasPermission/users.store
dependency into the package, the instanceOwner check is injected by
the caller via setIsInstanceOwner(). The editor-ui shim handles the
injection transparently so existing consumers keep working.

Changes:
- New @n8n/stores/src/cloudPlan.store.ts with injected instanceOwner
- New @n8n/stores/src/cloudPlan.store.test.ts (36 tests, all passing)
- editor-ui cloudPlan.store.ts replaced with deprecation shim
- Inlined CLOUD_TRIAL_CHECK_INTERVAL (sole consumer)
- Replaced luxon DateTime with native Date (drops dep)
- Typed catch blocks for strict mode
